Add AlarmClockView in addition to the AlarmClockLayout
